Create README.md
Title: Game Hub.
This Website is a place where you can buy games.
i  made  a  somme few changes  from  the design ca, and  this  is for  example not adding som symbols in the buy now buttons,  and i also did not include that many checkout pages, pluss  the CVV page. I stuck all of the card info and personal information in one page, to make it more clean and faster.
My references is  mostly youtube and google, and some fellow students.